+++ neo/renderer/draw_common.cpp
@@ -558,15 +558,17 @@ void RB_STD_FillDepthBuffer( drawSurf_t **drawSurfs, int numDrawSurfs ) {
 	RB_RenderDrawSurfListWithFunction( drawSurfs, numDrawSurfs, RB_T_FillDepthBuffer );
 
 	// Make the early depth pass available to shaders. #3877
-	if ( backEnd.viewDef->renderView.viewID >= 0  // Suppress for lightgem rendering passes
-		 && !r_skipDepthCapture.GetBool() )
+	bool getDepthCapture = r_enableDepthCapture.GetInteger() == 1
+		|| (r_enableDepthCapture.GetInteger() == -1 && r_useSoftParticles.GetBool());
+
+	if ( getDepthCapture && backEnd.viewDef->renderView.viewID >= 0 ) // Suppress for lightgem rendering passes
 	{
 		globalImages->currentDepthImage->CopyDepthbuffer( backEnd.viewDef->viewport.x1,
 														  backEnd.viewDef->viewport.y1,
 														  backEnd.viewDef->viewport.x2 - backEnd.viewDef->viewport.x1 + 1,
 														  backEnd.viewDef->viewport.y2 - backEnd.viewDef->viewport.y1 + 1, 
 														  true );
-		bool isPostProcess = false; // TODO
+		bool isPostProcess = false;
 		RB_SetProgramEnvironment( isPostProcess );
 	}